move Binary and Assign into expression module
