+++ boa_engine/src/bytecompiler/expression/assign.rs
@@ -1,0 +1,92 @@
+use boa_ast::expression::operator::{assign::AssignOp, Assign};
+
+use crate::{
+    bytecompiler::{Access, ByteCompiler},
+    vm::{BindingOpcode, Opcode},
+    JsResult,
+};
+
+pub(crate) fn compile_assign<'b>(
+    byte_compiler: &mut ByteCompiler<'b>,
+    assign: &Assign,
+    use_expr: bool,
+) -> JsResult<()> {
+    if assign.op() == AssignOp::Assign {
+        match Access::from_assign_target(assign.lhs()) {
+            Ok(access) => byte_compiler.access_set(access, use_expr, |compiler, _| {
+                compiler.compile_expr(assign.rhs(), true)?;
+                Ok(())
+            })?,
+            Err(pattern) => {
+                byte_compiler.compile_expr(assign.rhs(), true)?;
+                if use_expr {
+                    byte_compiler.emit_opcode(Opcode::Dup);
+                }
+                byte_compiler.compile_declaration_pattern(pattern, BindingOpcode::SetName)?;
+            }
+        }
+    } else {
+        let access = Access::from_assign_target(assign.lhs())
+            .expect("patterns should throw early errors on complex assignment operators");
+
+        let shortcircuit_operator_compilation =
+            |compiler: &mut ByteCompiler<'_>, opcode: Opcode| -> JsResult<()> {
+                let (early_exit, pop_count) =
+                    compiler.access_set(access, use_expr, |compiler, level| {
+                        compiler.access_get(access, true)?;
+                        let early_exit = compiler.emit_opcode_with_operand(opcode);
+                        compiler.compile_expr(assign.rhs(), true)?;
+                        Ok((early_exit, level))
+                    })?;
+                if pop_count == 0 {
+                    compiler.patch_jump(early_exit);
+                } else {
+                    let exit = compiler.emit_opcode_with_operand(Opcode::Jump);
+                    compiler.patch_jump(early_exit);
+                    for _ in 0..pop_count {
+                        compiler.emit_opcode(Opcode::Swap);
+                        compiler.emit_opcode(Opcode::Pop);
+                    }
+                    compiler.patch_jump(exit);
+                }
+                Ok(())
+            };
+
+        let opcode = match assign.op() {
+            AssignOp::Assign => unreachable!(),
+            AssignOp::Add => Opcode::Add,
+            AssignOp::Sub => Opcode::Sub,
+            AssignOp::Mul => Opcode::Mul,
+            AssignOp::Div => Opcode::Div,
+            AssignOp::Mod => Opcode::Mod,
+            AssignOp::Exp => Opcode::Pow,
+            AssignOp::And => Opcode::BitAnd,
+            AssignOp::Or => Opcode::BitOr,
+            AssignOp::Xor => Opcode::BitXor,
+            AssignOp::Shl => Opcode::ShiftLeft,
+            AssignOp::Shr => Opcode::ShiftRight,
+            AssignOp::Ushr => Opcode::UnsignedShiftRight,
+            AssignOp::BoolAnd => {
+                shortcircuit_operator_compilation(byte_compiler, Opcode::LogicalAnd)?;
+                return Ok(());
+            }
+            AssignOp::BoolOr => {
+                shortcircuit_operator_compilation(byte_compiler, Opcode::LogicalOr)?;
+                return Ok(());
+            }
+            AssignOp::Coalesce => {
+                shortcircuit_operator_compilation(byte_compiler, Opcode::Coalesce)?;
+                return Ok(());
+            }
+        };
+
+        byte_compiler.access_set(access, use_expr, |compiler, _| {
+            compiler.access_get(access, true)?;
+            compiler.compile_expr(assign.rhs(), true)?;
+            compiler.emit(opcode, &[]);
+            Ok(())
+        })?;
+    }
+
+    Ok(())
+}
