+++ boa_engine/src/bytecompiler/mod.rs
@@ -923,170 +923,8 @@ impl<'b> ByteCompiler<'b> {
         match expr {
             Expression::Literal(lit) => expression::compile_literal(self, lit, use_expr),
             Expression::Unary(unary) => expression::compile_unary(self, unary, use_expr)?,
-            Expression::Binary(binary) => {
-                self.compile_expr(binary.lhs(), true)?;
-                match binary.op() {
-                    BinaryOp::Arithmetic(op) => {
-                        self.compile_expr(binary.rhs(), true)?;
-                        match op {
-                            ArithmeticOp::Add => self.emit_opcode(Opcode::Add),
-                            ArithmeticOp::Sub => self.emit_opcode(Opcode::Sub),
-                            ArithmeticOp::Div => self.emit_opcode(Opcode::Div),
-                            ArithmeticOp::Mul => self.emit_opcode(Opcode::Mul),
-                            ArithmeticOp::Exp => self.emit_opcode(Opcode::Pow),
-                            ArithmeticOp::Mod => self.emit_opcode(Opcode::Mod),
-                        }
-
-                        if !use_expr {
-                            self.emit(Opcode::Pop, &[]);
-                        }
-                    }
-                    BinaryOp::Bitwise(op) => {
-                        self.compile_expr(binary.rhs(), true)?;
-                        match op {
-                            BitwiseOp::And => self.emit_opcode(Opcode::BitAnd),
-                            BitwiseOp::Or => self.emit_opcode(Opcode::BitOr),
-                            BitwiseOp::Xor => self.emit_opcode(Opcode::BitXor),
-                            BitwiseOp::Shl => self.emit_opcode(Opcode::ShiftLeft),
-                            BitwiseOp::Shr => self.emit_opcode(Opcode::ShiftRight),
-                            BitwiseOp::UShr => self.emit_opcode(Opcode::UnsignedShiftRight),
-                        }
-
-                        if !use_expr {
-                            self.emit(Opcode::Pop, &[]);
-                        }
-                    }
-                    BinaryOp::Relational(op) => {
-                        self.compile_expr(binary.rhs(), true)?;
-                        match op {
-                            RelationalOp::Equal => self.emit_opcode(Opcode::Eq),
-                            RelationalOp::NotEqual => self.emit_opcode(Opcode::NotEq),
-                            RelationalOp::StrictEqual => self.emit_opcode(Opcode::StrictEq),
-                            RelationalOp::StrictNotEqual => self.emit_opcode(Opcode::StrictNotEq),
-                            RelationalOp::GreaterThan => self.emit_opcode(Opcode::GreaterThan),
-                            RelationalOp::GreaterThanOrEqual => {
-                                self.emit_opcode(Opcode::GreaterThanOrEq);
-                            }
-                            RelationalOp::LessThan => self.emit_opcode(Opcode::LessThan),
-                            RelationalOp::LessThanOrEqual => self.emit_opcode(Opcode::LessThanOrEq),
-                            RelationalOp::In => self.emit_opcode(Opcode::In),
-                            RelationalOp::InstanceOf => self.emit_opcode(Opcode::InstanceOf),
-                        }
-
-                        if !use_expr {
-                            self.emit(Opcode::Pop, &[]);
-                        }
-                    }
-                    BinaryOp::Logical(op) => {
-                        match op {
-                            LogicalOp::And => {
-                                let exit = self.emit_opcode_with_operand(Opcode::LogicalAnd);
-                                self.compile_expr(binary.rhs(), true)?;
-                                self.patch_jump(exit);
-                            }
-                            LogicalOp::Or => {
-                                let exit = self.emit_opcode_with_operand(Opcode::LogicalOr);
-                                self.compile_expr(binary.rhs(), true)?;
-                                self.patch_jump(exit);
-                            }
-                            LogicalOp::Coalesce => {
-                                let exit = self.emit_opcode_with_operand(Opcode::Coalesce);
-                                self.compile_expr(binary.rhs(), true)?;
-                                self.patch_jump(exit);
-                            }
-                        };
-
-                        if !use_expr {
-                            self.emit(Opcode::Pop, &[]);
-                        }
-                    }
-                    BinaryOp::Comma => {
-                        self.emit(Opcode::Pop, &[]);
-                        self.compile_expr(binary.rhs(), true)?;
-
-                        if !use_expr {
-                            self.emit(Opcode::Pop, &[]);
-                        }
-                    }
-                }
-            }
-            Expression::Assign(assign) if assign.op() == AssignOp::Assign => {
-                match Access::from_assign_target(assign.lhs()) {
-                    Ok(access) => self.access_set(access, use_expr, |compiler, _| {
-                        compiler.compile_expr(assign.rhs(), true)?;
-                        Ok(())
-                    })?,
-                    Err(pattern) => {
-                        self.compile_expr(assign.rhs(), true)?;
-                        if use_expr {
-                            self.emit_opcode(Opcode::Dup);
-                        }
-                        self.compile_declaration_pattern(pattern, BindingOpcode::SetName)?;
-                    }
-                }
-            }
-            Expression::Assign(assign) => {
-                let access = Access::from_assign_target(assign.lhs())
-                    .expect("patterns should throw early errors on complex assignment operators");
-
-                let shortcircuit_operator_compilation =
-                    |compiler: &mut ByteCompiler<'_>, opcode: Opcode| -> JsResult<()> {
-                        let (early_exit, pop_count) =
-                            compiler.access_set(access, use_expr, |compiler, level| {
-                                compiler.access_get(access, true)?;
-                                let early_exit = compiler.emit_opcode_with_operand(opcode);
-                                compiler.compile_expr(assign.rhs(), true)?;
-                                Ok((early_exit, level))
-                            })?;
-                        if pop_count == 0 {
-                            compiler.patch_jump(early_exit);
-                        } else {
-                            let exit = compiler.emit_opcode_with_operand(Opcode::Jump);
-                            compiler.patch_jump(early_exit);
-                            for _ in 0..pop_count {
-                                compiler.emit_opcode(Opcode::Swap);
-                                compiler.emit_opcode(Opcode::Pop);
-                            }
-                            compiler.patch_jump(exit);
-                        }
-                        Ok(())
-                    };
-
-                let opcode = match assign.op() {
-                    AssignOp::Assign => unreachable!(),
-                    AssignOp::Add => Opcode::Add,
-                    AssignOp::Sub => Opcode::Sub,
-                    AssignOp::Mul => Opcode::Mul,
-                    AssignOp::Div => Opcode::Div,
-                    AssignOp::Mod => Opcode::Mod,
-                    AssignOp::Exp => Opcode::Pow,
-                    AssignOp::And => Opcode::BitAnd,
-                    AssignOp::Or => Opcode::BitOr,
-                    AssignOp::Xor => Opcode::BitXor,
-                    AssignOp::Shl => Opcode::ShiftLeft,
-                    AssignOp::Shr => Opcode::ShiftRight,
-                    AssignOp::Ushr => Opcode::UnsignedShiftRight,
-                    AssignOp::BoolAnd => {
-                        shortcircuit_operator_compilation(self, Opcode::LogicalAnd)?;
-                        return Ok(());
-                    }
-                    AssignOp::BoolOr => {
-                        shortcircuit_operator_compilation(self, Opcode::LogicalOr)?;
-                        return Ok(());
-                    }
-                    AssignOp::Coalesce => {
-                        shortcircuit_operator_compilation(self, Opcode::Coalesce)?;
-                        return Ok(());
-                    }
-                };
-
-                self.access_set(access, use_expr, |compiler, _| {
-                    compiler.access_get(access, true)?;
-                    compiler.compile_expr(assign.rhs(), true)?;
-                    compiler.emit(opcode, &[]);
-                    Ok(())
-                })?;
-            }
+            Expression::Binary(binary) => expression::compile_binary(self, binary, use_expr)?,
+            Expression::Assign(assign) => expression::compile_assign(self, assign, use_expr)?,
             Expression::ObjectLiteral(object) => {
                 self.emit_opcode(Opcode::PushEmptyObject);
                 for property in object.properties() {
